+++ example/data/settings/settings.dart
@@ -1,0 +1,3 @@
+class Settings {
+  final Duration waitTime = const Duration(milliseconds: 450);
+}