+++ example/data/user/model/user_model.dart
@@ -1,0 +1,7 @@
+class UserModel {
+  UserModel(this.id, this.name, this.age);
+
+  final String id;
+  final String name;
+  final String age;
+}