+++ example/data/user/user_data_source.dart
@@ -1,0 +1,5 @@
+import 'model/user_model.dart';
+
+abstract class UserDataSource {
+  Future<List<UserModel>> fetchUsers();
+}