+++ example/infrastructure/app.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+import 'routes.dart';
+
+MaterialApp get myApp => MaterialApp(
+      routes: routes,
+      initialRoute: initialRoute.path,
+    );