+++ example/infrastructure/base_router.dart
@@ -1,0 +1,7 @@
+import 'routes.dart';
+
+abstract class BaseRouter {
+  String get path;
+
+  RouteWidget get widget;
+}