+++ example/presentation/home/home_bloc.dart
@@ -1,0 +1,22 @@
+import 'dart:async';
+
+import '../../domain/user/get_user_use_case.dart';
+import '../../repository/user/model/user.dart';
+
+class HomeBloc {
+  HomeBloc(this._getUserUseCase) {
+    init();
+  }
+
+  final GetUserUseCase _getUserUseCase;
+
+  final StreamController<List<User>?> _userController = StreamController();
+
+  Stream<List<User>?> get userStream => _userController.stream;
+
+  Future<void> init() async {
+    _userController.sink.add(null);
+    final users = await _getUserUseCase.fetchOrderedUsers();
+    _userController.sink.add(users);
+  }
+}