+++ example/repository/user/model/user.dart
@@ -1,7 +1,7 @@
 class User {
   User(this.id, this.name, this.age);
 
-  final String id;
+  final int id;
   final String name;
   final int age;
 }