+++ example/repository/user/user_repository.dart
@@ -1,0 +1,5 @@
+import 'model/user.dart';
+
+abstract class UserRepository {
+  Future<List<User>> fetchUsers();
+}