+++ example/repository/user_repository.dart
@@ -1,5 +1,0 @@
-import '../model/user.dart';
-
-abstract class UserRepository {
-  Future<User> getUser();
-}