+++ lib/core/error/dependency_error.dart
@@ -1,6 +1,8 @@
+/// Wrapper class for [Error].
 class DependencyError extends Error {
   DependencyError(this.message);
 
+  /// Error message.
   final String message;
 
   @override