+++ lib/core/model/creation.dart
@@ -1,5 +1,0 @@
-abstract class Creation<T> {
-  abstract final Type type;
-
-  T get instance;
-}