+++ lib/core/model/factory.dart
@@ -1,15 +1,21 @@
-import 'package:dependon/core/model/creation.dart';
+import 'package:dependon/core/model/definition.dart';
+import 'package:dependon/core/utils/initialization.dart';
 
-class Factory<T> extends Creation<T> {
+/// A factory definition is an object that will be created every time `get` is
+/// called.
+class Factory<T> extends Definition<T> {
   Factory(this.initialization);
 
-  final Function() initialization;
+  /// Instruction for object initialization.
+  final Initialization initialization;
 
+  /// Retrieves a new instance of [T].
   @override
   T get instance {
     return initialization.call();
   }
 
+  /// Returns the primary [type] of [T].
   @override
   Type get type => T;
 }