+++ lib/core/utils/initialization.dart
@@ -1,0 +1,2 @@
+/// Lambda function for initialization.
+typedef Initialization<T> = T Function();