+++ lib/dependon.dart
@@ -1,7 +1,5 @@
 library dependon;
 
-export 'core/public/create_instance.dart';
-export 'core/public/factory.dart';
-export 'core/public/get.dart';
-export 'core/public/singleton.dart';
-export 'core/error/dependency_error.dart';
+export 'core/public/factory.dart' show factori;
+export 'core/public/get.dart' show get;
+export 'core/public/singleton.dart' show singleton;