+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.3
+
+- Removed createInstance method.
+- New feature: module
+- `factory` was renamed to `factori` in order to avoid conflicts with the Material's factory.
+- Added more code documentation.
+- Sample updated.
+
 ## 0.0.2
 
 - Removed dependency from collection's package.